+++ src/consts/usage.sh
@@ -2,13 +2,18 @@
 # shellcheck shell=bash
 readonly USAGE="\
 usage: $0 <file>...
-    Calls every passed down executable FILE and formats the output.
+    Add every FILE as a test.
+
+    Every test is executed and exit code, stdout and stderr output is saved and
+    displayed in a listed format.
 
     Options:
-      -c, --color=<when>  when to color the output (default: auto)
-                           WHEN may be: 'always', 'auto', or 'never'
-      -h, --help          show this summary and exit
-      -v, --version       show version and legal information and exit
+      -s, --silent=<file>  add FILE as test but don't save the output of it
+          --silent-all     don't save the output of any test
+      -c, --color=<when>   when to color the output (default: auto).
+                            WHEN may be: 'always', 'auto', or 'never'
+      -h, --help           show this summary and exit
+      -v, --version        show version and legal information and exit
 
     Exit Status:
       (using CommonCodes v1.0.0